+++ apps/www/src/api/models/fetchTextFromUpstream/responses.ts
@@ -3,7 +3,7 @@
 import { z } from "zod";
 import { c } from "~/ts-rest/root";
 import { createSchemaResponses, httpErrorSchema } from "~/@ashgw/ts-rest";
-import { internalErrorSchemaResponse } from "../_shared/responses";
+import { internalErrorSchemaResponse } from "../shared/responses";
 import type { InferResponses } from "~/@ashgw/ts-rest";
 
 // ========== Schemas ==========
