+++ apps/www/src/api/models/webhooks/purgeViewWindow/dtos.ts
@@ -1,7 +1,7 @@
 // TODO: to each schema we need to add as many docs as possible so the openAPI defintion
 // becomes rich so AI can use it
 import type { z } from "zod";
-import { cronAuthedMiddlewareHeaderSchemaDto } from "./../../_shared";
+import { cronAuthedMiddlewareHeaderSchemaDto } from "../../shared";
 
 // ========== Schemas ==========
