--- apps/www/src/ts-rest/middlewares/authed/index.ts
@@ -1,1 +1,0 @@
-export { cornAuthed } from "./cron";
